Added Sse2 fallback for Avx code
Added ReadOnlySpan<byte> Constructor, CopyTo and TryCopyTo implementations.

Fixed bug with indexer that allowed accessing values after the end of 64, 128 and 192 bit vectors.

Added Sse2 fallbacks for 192 and 256 bit Avx vector operations. This required access to the upper two doubles as a Vector128<double> in the 256 bit case which as before is done with Unsafe.As but now with a byte offset.