Added Register<T>
The physical storage of vectors should now be able to cope with any numeric value type, but this is untested currently. Double still works though.
The new functions at the bottom are a technique I saw in dotnet's Vector<T> struct and relies on the compiler to remove the massive if block in each and replace it with the literal in whatever case is right. This is the easiest method I could think of to vary implementation based on type without having a separate set of constructors for each.