Adding Vector<T> addition
Current implementation might be bad since it needs to get around intentional limitations in c# generic system twice. The x86 intrinsic operations only operate on vectors of concrete type and not the VectorX<T> used in this code. My current solution to this is to add an unsafe cast into the existing unsafe type conversion code to avoid a performance penalty. This change is shown in code by the two new Create functions, one for Vector<T> and one for Register<T> and changes to the Vector256<T> indexer to accept a concrete type at runtime. 

This results in addition working for 256 bit vectors whenever avx and avx2 are supported. The exact support required depends on the type since only float and double are supported by avx, the rest need avx2.