Add debug methods for Register
The new pointer array system is hard to debug as is so RegisterDebugView automatically converts it into managed arrays for each type. 

Additionally there is now DebugToString which displays the full vector in visual studio's debug window with the correct type. Not sure if the mess around required to get a type into that getter is worth it.