Unrolled arbitrary length addition code
Improved Register Create block overloads, should be slightly quicker now since operations are not repeated unnecessarily. 

Fixed unrolled 136 to 248 bit code for int and uint since indexes were wrong before. 
Most of the code in this and the previous commit is untested but ideally proper testing will come soon to deal with this. There are lots of cases so not sure how realistic it would be to test all of them.