Adding header file
Not done in an ideal manner currently as it turns out template classes are weird and don't work correctly when split between h and cpp files.

Also not totally sure about moving the string functions to another file, was mostly a test of having vector functions outside of the class which would be useful in cleaning up blendv.